# backend/engineapp/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Chess Engine App service responsible for game logic and AI moves.",
        "title": "EngineApp API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
        "/game/{gameId}": {
            "get": {
                "description": "retrieves the current state of a chess game from the database",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "games"
                ],
                "summary": "Get game state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "gameId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Game status",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "gameId is required or failed to fetch game",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health/live": {
            "get": {
                "description": "returns the health status of the engineapp service",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Check service health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "health.HealthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}